split configuration - TF 2.0 model working
