+++ pytorch_transformers/convert_gpt2_checkpoint_to_pytorch.py
@@ -21,10 +21,9 @@ from io import open
 
 import torch
 
-from pytorch_transformers.modeling_gpt2 import (CONFIG_NAME, WEIGHTS_NAME,
-                                                     GPT2Config,
-                                                     GPT2Model,
-                                                     load_tf_weights_in_gpt2)
+from pytorch_transformers import (CONFIG_NAME, WEIGHTS_NAME,
+                                  GPT2Config, GPT2Model,
+                                  load_tf_weights_in_gpt2)
 
 import logging
 logging.basicConfig(level=logging.INFO)
